Support the interval format for prohibited-days-dates (#37)
From this patch, `prohibited-days-dates` will support the interval
format as an [ISO 8610](https://en.wikipedia.org/wiki/ISO_8601#Time_intervals)
string. This would be convenient for users to specify the range of dates.

I also changed the type `Dates` to `Interval[]` in accordance with the
change because `Interval` in luxon is appropriate to describe the range.