Update README.md
The `synchronized` action seems to be unnecessary.

When the event with `synchronized` occurs, `opened` or `reopened` has already
happened. If `opened` or `reopened` appears during block time, the PR has
already been blocked. If `opened` or `reopened` appears during non-block time,
the PR has already been mergeable. If the time goes into block time between
`opened` or `reopened` and `synchronized`, the `schedule` event is supposed to
set the PR non-mergeable.