[Fix] Always run `createCommitStatus` at the pull request event
This tool should run when the action of pull request is `labeled` or
`unlabeled`, so it's not appropriate to see the list of commit statuses
for the head sha.

First, I made it check old commit statuses to avoid excessive
`createCommitStatus` API, but such cases might be edge cases. That's why
I ditched the feature with this patch.