Fix the test for inputs.prohibitedDates.length
The test for `inputs.prohibitedDates.length` is flaky because it depends
on the actual holidays of the specific region. Still, I want to check
whether the `prohibitedDates` are added by regional holidays, so I
change the matcher from `toEqual` to `toBeGreaterThan`. It's sufficient
to check the enough amount of prohibited days are added.