Fix __test__/inputs.ts
The test case `"returns a valid instance with base branches"` expects
the `prohibitedDates` has the twice number of prohibited days because `H:
` and `BH:` parameters. However, Spain's number of holiday has been
changed, so the test has failed.

We don't have to check the exact number of prohibited holidays, so the
patch set an enoufh days to the expected value.