Warn with unsupported events, instead of throwing
Sometimes, developers wants to use this GitHub Action in conjunction
with other jobs in one GitHub Workflow, and they might declare GitHub
events that are not supported by yykamei/block-merge-based-on-time.

Previously, this Action throwed an error when an unsupported event
triggers it, but I want to change the behavior not to raise an error.
Instead, I made changes to show warnings for developer experiences.