Add a new output, pr-blocked
If you want to do something based on the result of this action,
`pr-blocked` might be useful.

This new output will indicate the pull request is blocked or not
through an output parameter.

You can use it in your workflow like this:

```yaml
on:
  pull_request:
  schedule:
    - cron: "*/30 * * * *"
jobs:
  block:
    runs-on: ubuntu-latest
    steps:
      - uses: yykamei/block-merge-based-on-time@main
        id: block
        with:
          timezone: Asia/Tokyo
          after: 18:00
          before: 09:00
      - run: echo pr-blocked=${{ steps.block.outputs.pr-blocked }}
        if: github.event_name == 'pull_request'
```

Note `pr-blocked` won't be set on other than the `pull_request` event.