Use GitHub App token to run BMBT on forked PR (#2339)
Related to #1808

Currently, the workflow "Block Merge Based on Time" does not work if a
pull request is opened from a fork repository. This is because the
`GITHUB_TOKEN` doesn't have `write` permission to `statuses`.

See the following page for more details.

https://docs.github.com/en/actions/security-for-github-actions/security-guides/automatic-token-authentication#permissions-for-the-github_token

I want to update README to instruct how to allow the workflow to run on
pull requests from forked repositories.